+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Prune unused `./utils/` scripts, deps. Refs STCOR-959.
 * Don't show IST modal if `flsTimeRemaining` less than config.rtr.idleModalTTL. STCOR-884.
 * Allow customizing request timeout in `useOkapiKy`. STCOR-967.
+* Enhance locale management to support user and tenant locale configurations from both mod-configuration and mod-settings. STCOR-968.
 
 ## [11.0.0](https://github.com/folio-org/stripes-core/tree/v11.0.0) (2025-02-24)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v10.2.0...v11.0.0)